+++ src/components/Routes/MainScreen.jsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import Container from '../Layout/Container';
-import SideBar from '../Layout/SideBar';
 import Store from '../Layout/products/Store';
+import SideBar from '../Layout/sidebar/SideBar';
 import Window from '../Layout/Window';
 import WindowMain from '../Layout/WindowMain';
